It Works!
Is the biggest lie in computing. It does fail the tests... which is
probably expected, since there are new objects in play. But it does
"work"! Changes include frontend/search.js's cache stores subject
Name and subject count now, and the backend makes sure to always
check for subject, though honestly, it probably doesn't have to
anymore.